Refactor target process selection
Fixes #34

Refactor the process selection logic from `MainWindow` to a new `TargetSelectionWindow`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/theXappy/rnet-kit/issues/34?shareId=XXXX-XXXX-XXXX-XXXX).
